+++ Sandy Lodge/MHSystems/app/src/main/java/com/mh/systems/sandyLodge/adapter/RecyclerAdapter/ClubNewsSwipeAdapter.java
@@ -1,11 +1,9 @@
 package com.mh.systems.sandylodge.adapter.RecyclerAdapter;
 
-import android.app.Activity;
 import android.content.Context;
-
 import android.content.Intent;
+import android.os.Bundle;
 import android.support.v7.widget.RecyclerView;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,15 +14,12 @@ import android.widget.TextView;
 
 import com.daimajia.swipe.SwipeLayout;
 import com.daimajia.swipe.adapters.RecyclerSwipeAdapter;
-
 import com.mh.systems.sandylodge.R;
 import com.mh.systems.sandylodge.activites.ClubNewsActivity;
 import com.mh.systems.sandylodge.activites.ClubNewsDetailActivity;
 import com.mh.systems.sandylodge.models.ClubNews.ClubNewsData;
 
 import java.util.ArrayList;
-
-import static android.app.Activity.RESULT_OK;
 
 
 /**
@@ -208,13 +203,10 @@ public class ClubNewsSwipeAdapter extends RecyclerSwipeAdapter<ClubNewsSwipeAdap
         public void onClick(View itemView) {
 
             Intent detailNewsIntent = new Intent(mContext, ClubNewsDetailActivity.class);
-            detailNewsIntent.putExtra("ClubNewsID", clubNewsDataArrayList.get(getAdapterPosition()).getClubNewsID());
-            detailNewsIntent.putExtra("CreatedDate", clubNewsDataArrayList.get(getAdapterPosition()).getCreatedDate());
-            detailNewsIntent.putExtra("Time", clubNewsDataArrayList.get(getAdapterPosition()).getTime());
-            detailNewsIntent.putExtra("Message", clubNewsDataArrayList.get(getAdapterPosition()).getMessage());
-            detailNewsIntent.putExtra("IsRead", clubNewsDataArrayList.get(getAdapterPosition()).getIsRead());
-            detailNewsIntent.putExtra("iPosition", getAdapterPosition());
-
+            Bundle bundle = new Bundle();
+            bundle.putSerializable("club_news_content", clubNewsDataArrayList.get(getAdapterPosition()));
+            bundle.putInt("iPosition", getAdapterPosition());
+            detailNewsIntent.putExtras(bundle);
             ((ClubNewsActivity) mContext).startActivityForResult(detailNewsIntent, 111);
             //mContext.startActivity(detailNewsIntent);
         }