+++ Sunningdale/MHSystems/app/src/main/java/com/mh/systems/sunningdale/activites/ClubNewsDetailActivity.java
@@ -18,6 +18,7 @@ import com.mh.systems.sunningdale.R;
 import com.mh.systems.sunningdale.constants.ApplicationGlobal;
 import com.mh.systems.sunningdale.constants.WebAPI;
 import com.mh.systems.sunningdale.models.ClubNews.AJsonParamsClubNewsDetail;
+import com.mh.systems.sunningdale.models.ClubNews.ClubNewsData;
 import com.mh.systems.sunningdale.models.ClubNews.ClubNewsDetailAPI;
 import com.mh.systems.sunningdale.models.ClubNews.ClubNewsDetailResult;
 import com.mh.systems.sunningdale.util.API.WebServiceMethods;
@@ -60,6 +61,8 @@ public class ClubNewsDetailActivity extends BaseActivity {
 
     ClubNewsDetailResult clubNewsDetailResult;
 
+    ClubNewsData clubNewsData;
+
     /*********************************
      * DECLARATION OF CONSTANTS
      *******************************/
@@ -83,17 +86,21 @@ public class ClubNewsDetailActivity extends BaseActivity {
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         }
 
-        tvDateOfNews.setText(getIntent().getExtras().getString("CreatedDate"));
-        tvTimeOfNews.setText(getIntent().getExtras().getString("Time"));
-        // tvDescOfNews.setText(getIntent().getExtras().getString("Message"));
-        iClubNewsID = getIntent().getExtras().getInt("ClubNewsID");
-        isRead = getIntent().getExtras().getBoolean("IsRead");
+        Intent intent = this.getIntent();
+        Bundle bundle = intent.getExtras();
+        clubNewsData = (ClubNewsData) bundle.getSerializable("club_news_content");
 
-        iPosition = getIntent().getExtras().getInt("iPosition");
+        if (clubNewsData != null) {
+            tvDateOfNews.setText(clubNewsData.getCreatedDate());
+            tvTimeOfNews.setText(clubNewsData.getTime());
+            iClubNewsID = clubNewsData.getClubNewsID();
+            isRead = clubNewsData.getIsRead();
+            iPosition = bundle.getInt("iPosition");
+        }
 
         //data == html data which you want to load
         wvClubNews.getSettings().setJavaScriptEnabled(true);
-        wvClubNews.loadDataWithBaseURL("", getIntent().getExtras().getString("Message"), "text/html", "UTF-8", "");
+        wvClubNews.loadDataWithBaseURL("", clubNewsData.getMessage(), "text/html", "UTF-8", "");
 
         //If user haven't read news then call READ API status.
         if (!isRead) {