+++ Halesworth/MHSystems/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - New: Contact us feature added.
 - Fixed: Handicap graph date format 'MM/dd/yyyy' changed to 'dd/MM/yyyy'.
+- Fixed: Club News: Implements Serializable to pass club news data on detail screen instead of passing value one by one.
 
 ##Version 1.1.0 (2016-12-02)