+++ Porters Park/MHSystems/CHANGELOG.md
@@ -1,6 +1,7 @@
 ##Version  1.5.0 ()
 
 - Fixed: Handicap graph date format 'MM/dd/yyyy' changed to 'dd/MM/yyyy'.
+- Fixed: Club News: Implements Serializable to pass club news data on detail screen instead of passing value one by one.
 
 ##Version 1.4.0 (2016-12-02)